+++ src/actions/BitPayActions.js
@@ -155,7 +155,10 @@ export async function launchBitPay(
   const instructionOutput = invoiceInstruction.outputs[0]
 
   // Make the spend to generate the tx hexes
-  const requiredFeeRate = invoiceInstruction.requiredFeeRate
+  let requiredFeeRate = invoiceInstruction.requiredFeeRate
+  // This is in addition to the 1.5x multiplier in edge-currency-bitcoin. It's an additional buffer
+  // because the protocol doesn't discount segwit transactions and we want to make sure the transaction succeeds.
+  if (typeof requiredFeeRate === 'number') requiredFeeRate *= 1.2
   const spendInfo: EdgeSpendInfo = {
     selectedCurrencyCode,
     spendTargets: [
@@ -220,7 +223,6 @@ export async function launchBitPay(
       name: sprintf(s.strings.bitpay_metadata_name, paymentId),
       notes: sprintf(s.strings.bitpay_metadata_name, paymentId)
     },
-    dismissAlert: true,
     lockInputs: true,
     onDone: (error: Error | null, edgeTransaction?: EdgeTransaction) => {
       if (error) showError(`${s.strings.create_wallet_account_error_sending_transaction}: ${error.message}`)
@@ -229,11 +231,19 @@ export async function launchBitPay(
   }
 
   // Send confirmation scene
-  Actions.push('send', {
-    guiMakeSpendInfo,
-    selectedWalletId: selectedWallet.id,
-    selectedCurrencyCode
-  })
+  if (Actions.currentScene === 'send') {
+    Actions.refresh({
+      guiMakeSpendInfo,
+      selectedWalletId: selectedWallet?.id ?? '',
+      selectedCurrencyCode
+    })
+  } else {
+    Actions.push('send', {
+      guiMakeSpendInfo,
+      selectedWalletId: selectedWallet.id,
+      selectedCurrencyCode
+    })
+  }
 }
 
 /**