+++ src/components/modals/TransactionAdvanceDetails.js
@@ -2,7 +2,7 @@
 
 import type { EdgeTransaction } from 'edge-core-js'
 import React, { PureComponent } from 'react'
-import { Linking, Platform, ScrollView, StyleSheet, View } from 'react-native'
+import { Linking, Platform, ScrollView, View } from 'react-native'
 import { type AirshipBridge } from 'react-native-airship'
 import SafariView from 'react-native-safari-view'
 
@@ -74,32 +74,25 @@ class TransactionAdvanceDetailsComponent extends PureComponent<Props> {
     })
   }
 
-  renderFeeOptions(styles: StyleSheet) {
+  renderFeeOptions(): string {
     const { networkFeeOption, requestedCustomFee } = this.props.transaction
 
     if (networkFeeOption === 'custom') {
-      return this.renderFees(styles, s.strings.mining_fee_custom_label_choice, requestedCustomFee)
+      return `${s.strings.mining_fee_custom_label_choice}\n${this.renderFees(requestedCustomFee)}`
     }
-    return <EdgeText style={styles.text}>{networkFeeOption != null ? feeString[networkFeeOption] : s.strings.mining_fee_standard_label_choice}</EdgeText>
+
+    return networkFeeOption != null ? feeString[networkFeeOption] : s.strings.mining_fee_standard_label_choice
   }
 
-  renderFees(styles: StyleSheet, title: string, fees: Object = {}) {
-    const feeRows = []
+  renderFees(fees: Object = {}): string {
+    let feeValueText = ''
+
     for (const feeKey of Object.keys(fees)) {
-      const feeString = localizedFeeText[feeKey] ?? feeKey
-      feeRows.push(
-        <View key={feeKey} style={styles.feesRow}>
-          <EdgeText style={styles.feesRowText}>{feeString + ' '}</EdgeText>
-          <EdgeText style={styles.feesRowText}>{fees[feeKey]}</EdgeText>
-        </View>
-      )
+      const feeFullString = `${localizedFeeText[feeKey] ?? feeKey} ${fees[feeKey]}`
+      feeValueText = feeValueText === '' ? feeValueText + feeFullString : feeValueText + `\n${feeFullString}`
     }
-    return (
-      <View style={styles.feesContainer}>
-        <EdgeText style={styles.feesRowText}>{title + ':'}</EdgeText>
-        <View style={styles.feesBodyContainer}>{feeRows}</View>
-      </View>
-    )
+
+    return feeValueText
   }
 
   render() {
@@ -124,12 +117,10 @@ class TransactionAdvanceDetailsComponent extends PureComponent<Props> {
                 onPress={this.openUrl}
               />
             )}
-            {(networkFeeOption != null || feeRateUsed != null) && (
-              <Tile type="static" title={s.strings.transaction_details_advance_details_fee_info}>
-                {networkFeeOption != null ? this.renderFeeOptions(styles) : null}
-                {feeRateUsed != null ? this.renderFees(styles, s.strings.transaction_details_advance_details_fee_used, feeRateUsed) : null}
-              </Tile>
+            {networkFeeOption != null && (
+              <Tile type="static" title={s.strings.transaction_details_advance_details_fee_setting} body={this.renderFeeOptions()} />
             )}
+            {feeRateUsed != null && <Tile type="static" title={s.strings.transaction_details_advance_details_fee_used} body={this.renderFees(feeRateUsed)} />}
             {txSecret != null && <Tile type="copy" title={s.strings.transaction_details_advance_details_txSecret} body={txSecret} />}
             {txSecret != null && recipientAddress !== '' && txid !== '' && (
               <Tile
@@ -140,7 +131,7 @@ class TransactionAdvanceDetailsComponent extends PureComponent<Props> {
               />
             )}
             {signedTx != null && signedTx !== '' ? (
-              <Tile type="copy" title={s.strings.transaction_details_advance_details_raw_txbytes} body={signedTx} />
+              <Tile type="copy" title={s.strings.transaction_details_advance_details_raw_txbytes} body={signedTx} maximumHeight="small" />
             ) : null}
 
             {deviceDescription != null && <Tile type="static" title={s.strings.transaction_details_advance_details_device} body={deviceDescription} />}
@@ -169,25 +160,6 @@ const getStyles = cacheStyles((theme: Theme) => ({
   },
   body: {
     maxHeight: theme.rem(20)
-  },
-  text: {
-    color: theme.primaryText,
-    fontSize: theme.rem(1)
-  },
-  feesContainer: {
-    width: '100%',
-    flexDirection: 'row'
-  },
-  feesBodyContainer: {
-    flexDirection: 'column',
-    marginLeft: theme.rem(0.25)
-  },
-  feesRow: {
-    flexDirection: 'row'
-  },
-  feesRowText: {
-    color: theme.primaryText,
-    fontSize: theme.rem(1)
   }
 }))