+++ src/components/scenes/RequestScene.js
@@ -20,8 +20,8 @@ import { connect } from '../../types/reactRedux.js'
 import { type NavigationProp } from '../../types/routerTypes.js'
 import type { GuiCurrencyInfo, GuiDenomination, GuiWallet } from '../../types/types.js'
 import { getCurrencyIcon } from '../../util/CurrencyInfoHelpers.js'
-import { getWalletName } from '../../util/CurrencyWalletHelpers.js'
-import { getCurrencyInfo, getDenomFromIsoCode, getObjectDiff } from '../../util/utils.js'
+import { getAvailableBalance, getWalletName } from '../../util/CurrencyWalletHelpers.js'
+import { convertNativeToDenomination, getCurrencyInfo, getDenomFromIsoCode, getObjectDiff, truncateDecimals } from '../../util/utils.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 import { ButtonsModal } from '../modals/ButtonsModal.js'
 import { QrModal } from '../modals/QrModal.js'
@@ -31,6 +31,7 @@ import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services
 import { Card } from '../themed/Card.js'
 import { EdgeText } from '../themed/EdgeText.js'
 import { type ExchangedFlipInputAmounts, ExchangedFlipInput } from '../themed/ExchangedFlipInput.js'
+import { FiatText } from '../themed/FiatText.js'
 import { FlipInput } from '../themed/FlipInput.js'
 import { QrCode } from '../themed/QrCode'
 import { ShareButtons } from '../themed/ShareButtons.js'
@@ -277,10 +278,10 @@ export class RequestComponent extends React.Component<Props, State> {
   onError = (errorMessage?: string) => this.setState({ errorMessage })
 
   render() {
-    const { currencyIcon, exchangeSecondaryToPrimaryRatio, edgeWallet, primaryCurrencyInfo, secondaryCurrencyInfo, theme } = this.props
+    const { currencyIcon, exchangeSecondaryToPrimaryRatio, edgeWallet, guiWallet, primaryCurrencyInfo, secondaryCurrencyInfo, theme } = this.props
     const styles = getStyles(theme)
 
-    if (edgeWallet == null || primaryCurrencyInfo == null || secondaryCurrencyInfo == null || exchangeSecondaryToPrimaryRatio == null) {
+    if (guiWallet == null || primaryCurrencyInfo == null || secondaryCurrencyInfo == null || exchangeSecondaryToPrimaryRatio == null || edgeWallet == null) {
       return <ActivityIndicator color={theme.primaryText} style={styles.loader} size="large" />
     }
 
@@ -288,11 +289,35 @@ export class RequestComponent extends React.Component<Props, State> {
     const flipInputHeaderText = sprintf(s.strings.send_to_wallet, getWalletName(edgeWallet))
     const { keysOnlyMode = false } = getSpecialCurrencyInfo(edgeWallet.currencyInfo.pluginId)
 
+    // Balance
+    const nativeBalance = getAvailableBalance(edgeWallet, primaryCurrencyInfo.displayCurrencyCode)
+    const displayBalanceAmount = convertNativeToDenomination(primaryCurrencyInfo.displayDenomination.multiplier)(nativeBalance)
+    const displayBalanceString = sprintf(s.strings.request_balance, `${truncateDecimals(displayBalanceAmount)} ${primaryCurrencyInfo.displayDenomination.name}`)
+
+    // Selected denomination
+    const denomString = `1 ${primaryCurrencyInfo.displayDenomination.name}`
+
     return (
       <SceneWrapper background="header" hasTabs={false}>
         {keysOnlyMode !== true ? (
           <View style={styles.container}>
-            <EdgeText style={styles.title}>{s.strings.fragment_request_subtitle}</EdgeText>
+            <View style={styles.requestContainer}>
+              <EdgeText style={styles.title}>{s.strings.fragment_request_subtitle}</EdgeText>
+              <EdgeText style={styles.exchangeRate}>{denomString}</EdgeText>
+            </View>
+            <View style={styles.balanceContainer}>
+              <EdgeText>{displayBalanceString}</EdgeText>
+              <EdgeText style={styles.exchangeRate}>
+                <FiatText
+                  nativeCryptoAmount={primaryCurrencyInfo.displayDenomination.multiplier}
+                  cryptoCurrencyCode={primaryCurrencyInfo.displayCurrencyCode}
+                  isoFiatCurrencyCode={guiWallet.isoFiatCurrencyCode}
+                  autoPrecision
+                  appendFiatCurrencyCode
+                  cryptoExchangeMultiplier={primaryCurrencyInfo.exchangeDenomination.multiplier}
+                />
+              </EdgeText>
+            </View>
 
             {this.state.errorMessage != null ? <EdgeText style={styles.errorText}>{this.state.errorMessage}</EdgeText> : null}
 
@@ -497,11 +522,23 @@ const getStyles = cacheStyles((theme: Theme) => ({
     justifyContent: 'flex-start',
     paddingHorizontal: theme.rem(1)
   },
-
+  requestContainer: {
+    justifyContent: 'space-between',
+    flexDirection: 'row'
+  },
+  balanceContainer: {
+    justifyContent: 'space-between',
+    flexDirection: 'row',
+    marginBottom: theme.rem(0.5)
+  },
+  exchangeRate: {
+    textAlign: 'right',
+    alignSelf: 'flex-end',
+    fontFamily: theme.fontFaceBold
+  },
   title: {
     fontFamily: theme.fontFaceMedium,
-    fontSize: theme.rem(2),
-    marginBottom: theme.rem(0.5)
+    fontSize: theme.rem(2)
   },
 
   rightChevronContainer: {