+++ src/components/themed/AddressTile.js
@@ -124,9 +124,17 @@ class AddressTileComponent extends React.PureComponent<Props, State> {
     } catch (e) {
       const currencyInfo = coreWallet.currencyInfo
       const ercTokenStandard = currencyInfo.defaultSettings?.otherSettings?.ercTokenStandard ?? ''
-      if (ercTokenStandard === 'ERC20' && parseDeepLink(address).type === 'bitPay')
-        showError(new BitPayError('CurrencyNotSupported', { text: currencyInfo.currencyCode }))
-      else showError(`${s.strings.scan_invalid_address_error_title} ${s.strings.scan_invalid_address_error_description}`)
+      const parsedLink = { ...parseDeepLink(address) }
+      if (parsedLink.type === 'bitPay') {
+        if (ercTokenStandard === 'ERC20') {
+          showError(new BitPayError('CurrencyNotSupported', { text: currencyInfo.currencyCode }))
+        } else {
+          await launchBitPay(parsedLink.uri, { wallet: coreWallet }).catch(showError)
+        }
+      } else {
+        showError(`${s.strings.scan_invalid_address_error_title} ${s.strings.scan_invalid_address_error_description}`)
+      }
+
       this.setState({ loading: false })
     }
   }